BetaRowsetReader::next_block does not return 0 rows before eof
Doris uses both Status::Eof and 0 rows as eof in iterators, it is
a little messy. VCollectIterator::Level0Iterator::_refresh_current_row
returns EOF when empty block is returned, then VCollectIterator::build_heap
does not put the iterator to the heap.

So if we have a rowset of which the first batch rows is filtered by
pushed down predicates, then the rowset is not read any more. However,
there may be valid rows after the first batch.